socket api plumbing + deterboot
This commit introduces the minimal changes necessary to expose the udp
portion of the lwip sockets api to a syslinux module. Additionally it
shows the use of this exposed portion of the api through the deterboot
module. At the time of this commit, the deterboot module itself is
trivial and only verifies that sendto and recvfrom work as expected.

In the commits that follow:
- testing will be added for the lwip udp sockets api
- the deterboot module will be actually implemented

It is not our expectation that syslinux will incorporate the deterboot
module on the basis of its general applicability as it is only useful
within the context of the deter testbed. However, it may prove to be a
useful form of documentation by example for using the newly exposed
sockets api.