first bootinfo protocol test
This commit adds the first protocol test for bootinfo from the client
side. The test assumes that the server side has a configuration that
will result in a BOOTWAIT response. The necessary testing setup code to
accomplish this has been added to the testbed repository under
/test/boot0.

There was a bit of further sockets plumbing that had to be done to
accomodate the bootinfo protocol. Because this protocol uses different
ports for the senders and recievers, it was necessary to expose the
lwip_bind function so we could bind to the client side bootinfo
receiving port.

This commit also introduces a question/ask api for the deterboot client
to talk to a bootinfo server. This significanlty simplifies client code.
See the deterboot-test.c source to see how it works.