boot mfs test passing in deter env
This commit completes a basic test of the deterboot loader's ability to
load the linux-mfs using the bootinfo protocol with a boss node. This
test is successfully running in the vbed environment.

I am phasing out the old testing.c/testing.h local testing strategy in
favor of walrustf, this phasing out will take place over the next few
commits.